Move fetch and push options to egit-options.c
