+++ src/egit-remote.c
@@ -3,397 +3,10 @@
 #include "git2.h"
 
 #include "egit.h"
+#include "egit-options.h"
 #include "egit-util.h"
 #include "interface.h"
 #include "egit-remote.h"
-
-
-// =============================================================================
-// Callbacks
-
-typedef struct {
-    emacs_env *env;
-    emacs_value sideband_progress;
-    emacs_value certificate_check;
-    emacs_value credentials;
-    emacs_value transfer_progress;
-} remote_ctx;
-
-static int sideband_progress_cb(const char *str, int len, void *payload)
-{
-    remote_ctx *ctx = (remote_ctx*) payload;
-    emacs_env *env = ctx->env;
-
-    emacs_value msg = env->make_string(env, str, len);
-    env->funcall(env, ctx->sideband_progress, 1, &msg);
-
-    EM_RETURN_IF_NLE(GIT_EUSER);
-    return 0;
-}
-
-static int certificate_check_cb(git_cert *cert, int valid, const char *host, void *payload)
-{
-    remote_ctx *ctx = (remote_ctx*) payload;
-    emacs_env *env = ctx->env;
-
-    emacs_value args[3];
-    args[0] = em_nil;
-    args[1] = valid ? em_t : em_nil;
-    args[2] = host ? EM_STRING(host) : em_nil;
-
-    // TODO: GIT_CERT_STRARRAY?
-    if (cert->cert_type == GIT_CERT_X509) {
-        git_cert_x509 *c = (git_cert_x509*) cert;
-        emacs_value data = env->make_string(env, c->data, c->len);
-        data = em_string_as_unibyte(env, data);
-        args[0] = em_cons(env, em_x509, em_cons(env, data, em_nil));
-    }
-    else if (cert->cert_type == GIT_CERT_HOSTKEY_LIBSSH2) {
-        git_cert_hostkey *c = (git_cert_hostkey*) cert;
-
-        size_t nelts = 0;
-        emacs_value elts[5];
-        elts[nelts++] = em_hostkey_libssh2;
-
-        // We use libgit's own oid_tostr to decode hashes into hex strings
-        char buf[41];
-        if (c->type & GIT_CERT_SSH_MD5) {
-            elts[nelts++] = em_md5;
-            git_oid_tostr(buf, 33, (git_oid*) c->hash_md5);
-            elts[nelts++] = EM_STRING(buf);
-        }
-        if (c->type & GIT_CERT_SSH_SHA1) {
-            elts[nelts++] = em_sha1;
-            git_oid_tostr(buf, 41, (git_oid*) c->hash_sha1);
-            elts[nelts++] = EM_STRING(buf);
-        }
-
-        args[0] = em_list(env, elts, nelts);
-    }
-
-    env->funcall(env, ctx->certificate_check, 3, args);
-    EM_RETURN_IF_NLE(GIT_EUSER);
-    return 0;
-}
-
-static int credentials_cb(
-    git_cred **cred, const char *url, const char *username_from_url,
-    unsigned int allowed_types, void *payload)
-{
-    remote_ctx *ctx = (remote_ctx*) payload;
-    emacs_env *env = ctx->env;
-
-    emacs_value args[3];
-    args[0] = EM_STRING(url);
-    args[1] = username_from_url ? EM_STRING(username_from_url) : em_nil;
-
-    emacs_value types[7];
-    size_t ntypes = 0;
-    if (allowed_types & GIT_CREDTYPE_USERPASS_PLAINTEXT)
-        types[ntypes++] = em_userpass_plaintext;
-    if (allowed_types & GIT_CREDTYPE_SSH_KEY)
-        types[ntypes++] = em_ssh_key;
-    if (allowed_types & GIT_CREDTYPE_SSH_CUSTOM)
-        types[ntypes++] = em_ssh_custom;
-    if (allowed_types & GIT_CREDTYPE_DEFAULT)
-        types[ntypes++] = em_default;
-    if (allowed_types & GIT_CREDTYPE_SSH_INTERACTIVE)
-        types[ntypes++] = em_ssh_interactive;
-    if (allowed_types & GIT_CREDTYPE_USERNAME)
-        types[ntypes++] = em_username;
-    if (allowed_types & GIT_CREDTYPE_SSH_MEMORY)
-        types[ntypes++] = em_ssh_memory;
-    args[2] = em_list(env, types, ntypes);
-
-    emacs_value retval = env->funcall(env, ctx->credentials, 3, args);
-    EM_RETURN_IF_NLE(GIT_EUSER);
-
-    if (egit_get_type(env, retval) != EGIT_CRED) {
-        em_signal_wrong_type(env, em_libgit_cred_p, retval);
-        return GIT_EUSER;
-    }
-
-    // The transport object takes ownership of the credential,
-    // so we must duplicate it here.
-    return egit_cred_dup(cred, EGIT_EXTRACT(retval));
-}
-
-static int transfer_progress_cb(const git_transfer_progress *stats, void *payload)
-{
-    remote_ctx *ctx = (remote_ctx*) payload;
-    emacs_env *env = ctx->env;
-
-    emacs_value args[7];
-    args[0] = EM_INTEGER(stats->total_objects);
-    args[1] = EM_INTEGER(stats->indexed_objects);
-    args[2] = EM_INTEGER(stats->received_objects);
-    args[3] = EM_INTEGER(stats->local_objects);
-    args[4] = EM_INTEGER(stats->total_deltas);
-    args[5] = EM_INTEGER(stats->indexed_deltas);
-    args[6] = EM_INTEGER(stats->received_bytes);
-
-    env->funcall(env, ctx->transfer_progress, 7, args);
-    EM_RETURN_IF_NLE(GIT_EUSER);
-    return 0;
-}
-
-
-// =============================================================================
-// Helpers - option parsing
-
-static emacs_value proxy_parse(emacs_env *env, emacs_value alist, git_proxy_options *opts)
-{
-    emacs_value url = em_nil;
-    emacs_value cred = em_nil;
-    emacs_value cert = em_nil;
-
-    emacs_value car, cdr;
-    {
-        EM_DOLIST(cell, alist, options);
-        EM_ASSERT_CONS(cell);
-        car = em_car(env, cell);
-        cdr = em_cdr(env, cell);
-
-        if (EM_EQ(car, em_type)) {
-            if (!EM_EXTRACT_BOOLEAN(cdr))
-                opts->type = GIT_PROXY_NONE;
-            else if (EM_EQ(cdr, em_auto))
-                opts->type = GIT_PROXY_AUTO;
-            else if (EM_EQ(cdr, em_specified))
-                opts->type = GIT_PROXY_SPECIFIED;
-            else {
-                em_signal_wrong_value(env, cdr);
-                return em_nil;
-            }
-        }
-        else if (EM_EQ(car, em_url)) {
-            EM_ASSERT_STRING(cdr);
-            url = cdr;
-        }
-        else if (EM_EQ(car, em_credentials)) {
-            EM_ASSERT_FUNCTION(cdr);
-            cred = cdr;
-        }
-        else if (EM_EQ(car, em_certificate_check)) {
-            EM_ASSERT_FUNCTION(cdr);
-            cert = cdr;
-        }
-
-        EM_DOLIST_END(options);
-    }
-
-    if (EM_EXTRACT_BOOLEAN(url))
-        opts->url = EM_EXTRACT_STRING(url);
-    if (EM_EXTRACT_BOOLEAN(cred) || EM_EXTRACT_BOOLEAN(cert)) {
-        remote_ctx *payload = (remote_ctx*) malloc(sizeof(remote_ctx));
-        *payload = (remote_ctx) {.env = env, .credentials = cred, .certificate_check = cert};
-        if (EM_EXTRACT_BOOLEAN(cred))
-            opts->credentials = &credentials_cb;
-        if (EM_EXTRACT_BOOLEAN(cert))
-            opts->certificate_check = &certificate_check_cb;
-    }
-
-    return em_nil;
-}
-
-#define CHECK(sym, var)                         \
-    if (EM_EQ(car, (sym))) {                    \
-        EM_ASSERT_FUNCTION(cdr);                \
-        (var) = cdr;                            \
-        found = true;                           \
-    }
-
-#define STORE(sym, var)                         \
-    if (EM_EXTRACT_BOOLEAN((var))) {            \
-        opts->sym = &sym##_cb;                  \
-        payload->sym = (var);                   \
-    }
-
-static emacs_value callbacks_parse(emacs_env *env, emacs_value alist, git_remote_callbacks *opts)
-{
-    emacs_value side = em_nil,
-                cred = em_nil,
-                cert = em_nil,
-                prog = em_nil;
-    bool found = false;
-
-    emacs_value car, cdr;
-    {
-        EM_DOLIST(cell, alist, options);
-        EM_ASSERT_CONS(cell);
-        car = em_car(env, cell);
-        cdr = em_cdr(env, cell);
-
-        CHECK(em_sideband_progress, side);
-        CHECK(em_credentials, cred);
-        CHECK(em_certificate_check, cert);
-        CHECK(em_transfer_progress, prog);
-
-        EM_DOLIST_END(options);
-    }
-
-    if (found) {
-        remote_ctx *payload = (remote_ctx*) malloc(sizeof(remote_ctx));
-        *payload = (remote_ctx) {.env = env};
-        opts->payload = payload;
-        STORE(sideband_progress, side);
-        STORE(credentials, cred);
-        STORE(certificate_check, cert);
-        STORE(transfer_progress, prog);
-    }
-
-    return em_nil;
-}
-
-#undef CHECK
-#undef STORE
-
-static void fetch_options_dispose(git_fetch_options *opts)
-{
-    egit_strarray_dispose(&opts->custom_headers);
-    free(opts->proxy_opts.payload);
-    free((void*) opts->proxy_opts.url);
-    free(opts->callbacks.payload);
-}
-
-static emacs_value fetch_options_parse(emacs_env *env, emacs_value alist, git_fetch_options *opts)
-{
-    int retval = git_fetch_init_options(opts, GIT_FETCH_OPTIONS_VERSION);
-    EGIT_CHECK_ERROR(retval);
-
-    emacs_value callbacks = em_nil;
-    emacs_value headers = em_nil;
-    emacs_value proxy = em_nil;
-
-    // Main loop through the options alist
-    emacs_value car, cdr;
-    {
-        EM_DOLIST(cell, alist, options);
-        car = em_car(env, cell);
-        cdr = em_cdr(env, cell);
-
-        if (EM_EQ(car, em_callbacks))
-            callbacks = cdr;
-        else if (EM_EQ(car, em_headers))
-            headers = cdr;
-        else if (EM_EQ(car, em_proxy))
-            proxy = cdr;
-        else if (EM_EQ(car, em_prune)) {
-            if (!EM_EXTRACT_BOOLEAN(cdr))
-                opts->prune = GIT_FETCH_PRUNE_UNSPECIFIED;
-            else if (EM_EQ(cdr, em_on))
-                opts->prune = GIT_FETCH_PRUNE;
-            else if (EM_EQ(cdr, em_off))
-                opts->prune = GIT_FETCH_NO_PRUNE;
-            else {
-                em_signal_wrong_value(env, cdr);
-                return em_nil;
-            }
-        }
-        else if (EM_EQ(car, em_download_tags)) {
-            if (!EM_EXTRACT_BOOLEAN(cdr))
-                opts->download_tags = GIT_REMOTE_DOWNLOAD_TAGS_UNSPECIFIED;
-            else if (EM_EQ(cdr, em_auto))
-                opts->download_tags = GIT_REMOTE_DOWNLOAD_TAGS_AUTO;
-            else if (EM_EQ(cdr, em_none))
-                opts->download_tags = GIT_REMOTE_DOWNLOAD_TAGS_NONE;
-            else if (EM_EQ(cdr, em_all))
-                opts->download_tags = GIT_REMOTE_DOWNLOAD_TAGS_ALL;
-            else {
-                em_signal_wrong_value(env, cdr);
-                return em_nil;
-            }
-        }
-        else if (EM_EQ(car, em_update_fetchhead))
-            opts->update_fetchhead = EM_EXTRACT_BOOLEAN(cdr);
-
-        EM_DOLIST_END(options);
-    }
-
-    if (EM_EXTRACT_BOOLEAN(callbacks)) {
-        callbacks_parse(env, callbacks, &opts->callbacks);
-        if (env->non_local_exit_check(env))
-            goto cleanup;
-    }
-    if (EM_EXTRACT_BOOLEAN(headers)) {
-        if (!egit_strarray_from_list(&opts->custom_headers, env, headers))
-            goto cleanup;
-    }
-    if (EM_EXTRACT_BOOLEAN(proxy)) {
-        proxy_parse(env, proxy, &opts->proxy_opts);
-        if (env->non_local_exit_check(env))
-            goto cleanup;
-    }
-
-    return em_nil;
-
-  cleanup:
-    fetch_options_dispose(opts);
-    return em_nil;
-}
-
-static void push_options_dispose(git_push_options *opts)
-{
-    egit_strarray_dispose(&opts->custom_headers);
-    free(opts->proxy_opts.payload);
-    free((void*) opts->proxy_opts.url);
-    free(opts->callbacks.payload);
-}
-
-static emacs_value push_options_parse(emacs_env *env, emacs_value alist, git_push_options *opts)
-{
-    int retval = git_push_init_options(opts, GIT_PUSH_OPTIONS_VERSION);
-    EGIT_CHECK_ERROR(retval);
-
-    emacs_value callbacks = em_nil;
-    emacs_value headers = em_nil;
-    emacs_value proxy = em_nil;
-
-    // Main loop through the options alist
-    emacs_value car, cdr;
-    {
-        EM_DOLIST(cell, alist, options);
-        EM_ASSERT_CONS(cell);
-        car = em_car(env, cell);
-        cdr = em_cdr(env, cell);
-
-        if (EM_EQ(car, em_callbacks))
-            callbacks = cdr;
-        else if (EM_EQ(car, em_headers))
-            headers = cdr;
-        else if (EM_EQ(car, em_proxy))
-            proxy = cdr;
-        else if (EM_EQ(car, em_threads)) {
-            if (!EM_EXTRACT_BOOLEAN(cdr))
-                opts->pb_parallelism = 0;
-            EM_ASSERT_INTEGER(cdr);
-            opts->pb_parallelism = EM_EXTRACT_INTEGER(cdr);
-        }
-
-        EM_DOLIST_END(options);
-    }
-
-    if (EM_EXTRACT_BOOLEAN(callbacks)) {
-        callbacks_parse(env, callbacks, &opts->callbacks);
-        if (env->non_local_exit_check(env))
-            goto cleanup;
-    }
-    if (EM_EXTRACT_BOOLEAN(headers)) {
-        if (!egit_strarray_from_list(&opts->custom_headers, env, headers))
-            goto cleanup;
-    }
-    if (EM_EXTRACT_BOOLEAN(proxy)) {
-        proxy_parse(env, proxy, &opts->proxy_opts);
-        if (env->non_local_exit_check(env))
-            goto cleanup;
-    }
-
-    return em_nil;
-
-  cleanup:
-    push_options_dispose(opts);
-    return em_nil;
-}
 
 
 // =============================================================================
@@ -656,7 +269,7 @@ emacs_value egit_remote_fetch(
         return em_nil;
 
     git_fetch_options options;
-    fetch_options_parse(env, opts, &options);
+    egit_fetch_options_parse(env, opts, &options);
     if (env->non_local_exit_check(env)) {
         egit_strarray_dispose(&refspecs);
         return em_nil;
@@ -668,7 +281,7 @@ emacs_value egit_remote_fetch(
 
     free(msg);
     egit_strarray_dispose(&refspecs);
-    fetch_options_dispose(&options);
+    egit_fetch_options_release(&options);
 
     EGIT_CHECK_ERROR(retval);
     return em_nil;
@@ -692,7 +305,7 @@ emacs_value egit_remote_push(emacs_env *env, emacs_value _remote, emacs_value _r
         return em_nil;
 
     git_push_options options;
-    push_options_parse(env, opts, &options);
+    egit_push_options_parse(env, opts, &options);
     if (env->non_local_exit_check(env)) {
         egit_strarray_dispose(&refspecs);
         return em_nil;
@@ -702,7 +315,7 @@ emacs_value egit_remote_push(emacs_env *env, emacs_value _remote, emacs_value _r
     int retval = git_remote_push(remote, &refspecs, &options);
 
     egit_strarray_dispose(&refspecs);
-    push_options_dispose(&options);
+    egit_push_options_release(&options);
 
     EGIT_CHECK_ERROR(retval);
     return em_nil;
